@dhiAmzn 2021-04-26
In favor of https://github.com/opensearch-project/security/pull/1135